+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Added `deadgrep-project-root-overrides`. This enables you to change
 the root directory for a specific project without writing a whole custom
 function for `deadgrep-project-root-function`.
 
+Search results in narrowed buffers are now handled correctly.
+
 # v0.6
 
 `next-error` and `previous-error` now work, just like they do with